Refactor Socrata module
