+++ lib/tasks/socrata.rake
@@ -6,14 +6,14 @@ namespace :socrata do
   # - format - The exporter format as a string. Defaults to "CSV"
   # - i.e. rake socrata:export["JSON"]
   task :export, [:format] => :environment do |_task, args|
-    file_name, export_data = Socrata::Exporter.export(args.format)
+    file_name, export_data = Socrata.export(args.format)
     File.write(file_name, export_data)
   end
 
-  desc 'Updates Socrata dataset via HTTP POST request'
-  # The response object will be a summary of the updates were made
+  desc 'Updates the Socrata dataset via HTTP POST request'
+  # The response object will be a summary of the updates that were made
   # and possibly the errors that were encountered.
   task publish: :environment do
-    Socrata::Publisher.publish
+    Socrata.publish
   end
 end
